RENAME: change location of down_audio_from_bucket
